+++ test/unittest/random.jl
@@ -9,11 +9,11 @@ function test_uniform()
   low = -10; high = 10
   seed = 123
   mx.srand(seed)
-  ret1 = mx.rand(low, high, dims)
+  ret1 = mx.rand(dims..., low = low, high = high)
 
   mx.srand(seed)
   ret2 = mx.empty(dims)
-  mx.rand!(low, high, ret2)
+  mx.rand!(ret2, low = low, high = high)
 
   @test copy(ret1) == copy(ret2)
   @test abs(mean(copy(ret1)) - (high+low)/2) < 0.1
@@ -26,11 +26,11 @@ function test_gaussian()
   μ = 10; σ = 2
   seed = 456
   mx.srand(seed)
-  ret1 = mx.randn(μ, σ, dims)
+  ret1 = mx.randn(dims..., μ = μ, σ = σ)
 
   mx.srand(seed)
   ret2 = mx.empty(dims)
-  mx.randn!(μ, σ, ret2)
+  mx.randn!(ret2, μ = μ, σ = σ)
 
   @test copy(ret1) == copy(ret2)
   @test abs(mean(copy(ret1)) - μ) < 0.1